requirements.txt
pyzbar==0.1.8
Pillow==7.2.0
PyQt5==5.15.0
